pyproject.toml
[project]
name = "liteai"
version = "0.2.1"
description = "lite ai client"
requires-python = ">=3.10"
dependencies = [
    "python-snippets>=0.1.3",
    "python-dotenv>=1.0.1",
    # llm api
    "openai>=1.4.1",
    "zhipuai>=2.1.4.20230812",
    "dashscope>=1.20.0",
    "volcengine-python-sdk[ark]>=1.0.94",
    "litellm>=1.49.1",

    # audio and image
    "pillow>=10.3.0",
    "pydub>=0.25.1",


]
authors = [{ name = "jerry", email = "jerrychen1990@gmail.com" }]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
[tool.hatch.build.targets.wheel]
packages = ["liteai"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 140 # 代码最大行宽
select = [ # 选择的规则
    "F",
    "E",
    "W",
    "UP",
]
ignore = ["F401", "F405", "UP009", "F403", "F541"] # 忽略的规则
